requirement.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
asn1crypto==1.4.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
contextvars==2.4
cryptography==2.1.4
dataclasses==0.8
Django==3.2.4
django-cors-headers==2.4.0
django-debug-toolbar==3.2.1
django-pydash-app==1.4.4
django-redis==4.10.0
django-simpleui==2021.4
djangorestframework==3.11.1
Flask==2.0.1
gevent==21.1.2
greenlet==1.1.0
grequests==0.6.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.3
httpx==0.18.1
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
mysqlclient==2.0.3
nest-asyncio==1.5.1
pycparser==2.20
pytz==2021.1
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
